// apps/frontend/packages/api/base/codegen/types/FactPackageCfoStatus2.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Cfo } from "./Cfo";
import type { FactCfoStatusValue2 } from "./FactCfoStatusValue2";
import type { UserSummary } from "./UserSummary";

export interface FactPackageCfoStatus2 {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type integer
    */
    factPackageId: number;
    /**
     * @type integer
    */
    cfoId: number;
    /**
     * @type object
    */
    cfo: Cfo;
    /**
     * @type string
    */
    status: FactCfoStatusValue2;
    /**
     * @type integer,null
    */
    decidedById?: number | null;
    decidedBy?: (UserSummary | null);
    /**
     * @type string,null, date-time
    */
    decidedAt?: string | null;
}